+++ wemake_python_styleguide/compat/aliases.py
@@ -7,10 +7,24 @@ And provide links to the python source code.
 """
 
 import ast
-from typing import Final
+from typing import Final, final
 
-#: We need this tuple to easily work with both types of text nodes:
-TextNodes: Final = (ast.Str, ast.Bytes)
+
+@final
+class _TextNodesMeta(type):
+    def __instancecheck__(cls, instance):
+        return isinstance(instance, ast.Constant) and isinstance(
+            instance.value,
+            str | bytes,
+        )
+
+
+@final
+class TextNodes(ast.AST, metaclass=_TextNodesMeta):
+    """Check if node has type of `ast.Constant` with `str` or `bytes`."""
+
+    value: str | bytes  # noqa: WPS110
+
 
 #: We need this tuple to easily check that this is a real assign node.
 AssignNodes: Final = (ast.Assign, ast.AnnAssign)