+++ wemake_python_styleguide/visitors/ast/classes.py
@@ -334,15 +334,17 @@ class WrongSlotsVisitor(base.BaseNodeVisitor):
             self._count_slots_items(node, node.value)
 
     def _slot_item_name(self, node: ast.AST) -> str | None:
-        if isinstance(node, ast.Str):
-            return node.s
+        if isinstance(node, ast.Constant) and isinstance(node.value, str):
+            return node.value
         if isinstance(node, ast.Starred):
             return source.node_to_string(node)
         return None
 
     def _are_correct_slots(self, slots: list[ast.AST]) -> bool:
         return all(
-            slot.s.isidentifier() for slot in slots if isinstance(slot, ast.Str)
+            slot.value.isidentifier()
+            for slot in slots
+            if isinstance(slot, ast.Constant) and isinstance(slot.value, str)
         )